requirements.txt
pyping==0.0.6
python-json-logger==0.1.11
pyinstaller==3.6
flask 
flask_classful
waitress 
werkzeug
pytest
pyyaml 
pillow 
werkzeug
waitress
pathlib 